+++ CHANGELOG.md
@@ -9,6 +9,12 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
+## [Unreleased]
+
+## Changed
+
+- `NameOf()` and `TypeOf()` now produce a more meaningful panic message when passed a `nil` message
+
 ## [0.2.0] - 2019-12-16
 
 ### Removed